renamed and refactored a few things
